+++ can_ethernet.h
@@ -5,8 +5,8 @@
 #define DHCP_EXPIRE_TIMER_SECS 45			// Timeout for DHCP address request
 #endif
 
-void display_ip_address(unsigned long ipaddr, unsigned long ulCol,unsigned long ulRow);									// display an lwIP address
-void display_can_statistics(unsigned long msg_count, unsigned long lost_count, unsigned long col, unsigned long row);	// display CAN message statistics
+void display_ip_address(uint32_t ipaddr, uint32_t ulCol,uint32_t ulRow);									// display an lwIP address
+void display_can_statistics(uint32_t msg_count, uint32_t lost_count, uint32_t col, uint32_t row);	// display CAN message statistics
 void lwIPHostTimerHandler(void);																						// This function is required by lwIP library to support any host-related timer functions.
 
 #endif
